update component userView for UserProfile
