+++ src/styles/components/Buttonaction.scss
@@ -167,14 +167,40 @@
   }
 }
 
-.btn-action--videoup {
+/* ----------------------------------------------------- */
+/* BUTTON PROFILE STYLES */
+/* ----------------------------------------------------- */
+
+.btn-action--profile {
   display: flex;
   flex-grow: 1;
   text-align: center;
   justify-content: center;
+  border: none;
+  border-radius: 0.25rem;
   background-color: #065fd4;
   color: white;
-  border: none;
+  text-transform: uppercase;
+  font-size: 0.8rem;
+  font-weight: 700;
+  transition-property: opacity;
+  transition-duration: 200ms;
+
+  @include tabletAndDesktop() {
+    font-size: 1rem;
+  }
+}
+
+.btn-action--profile--delete {
+  border-style: solid;
+  border-width: 1px;
+  border-color: var(--red);
+  background-color: var(--light);
+  color: var(--red);
+}
+
+.btn-action--profile:disabled {
+  opacity: 0.5;
 }
 
 .btn-action--tabsContainer {
