+++ src/styles/pages/UserProfile.scss
@@ -1,6 +1,7 @@
-.page {
+.user-profile {
   margin-top: 5px;
   margin-left: 5px;
+  margin-right: 10px;
 }
 
 .page__headerview {
